pyproject.toml
[tool.poetry]
name = "imagepackages"
version = "0.1.0"
description = ""
authors = ["Roman <roman.nieronov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.1"
pillow = "^10.0.0"
torchvision = "^0.15.2"
openai-clip = "^1.0.1"
pytesseract = "^0.3.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
